% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute}
\alias{execute}
\title{Execute imap request.}
\usage{
execute(imap, filename = NULL, ...)
}
\arguments{
\item{imap}{imap object}

\item{...}{optional parameters passed to curl
like \code{ssl_verifypeer = F}
(see \link[curl]{curl_options})}
}
\value{
list containing url of request
   and content
}
\description{
Execute imap request.
}
\examples{
\dontrun{
   imap <- IMAP("imaps://imap.gmail.com") \%>\%
      user("name", "pwd")
   response <- imap \%>\% LIST() \%>\% execute()
   cat(response$content)
}

}

