% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_obs}
\alias{execute_obs}
\title{execute}
\usage{
execute_obs(imap)
}
\arguments{
\item{imap}{imap object}
}
\value{
response (list containing url, content and headers)
}
\description{
Execute imap request.
}
\examples{
\dontrun{
   imap <- IMAP("imaps://imap.gmail.com") \%>\%
      user("name", "pwd")
   response <- imap \%>\% LIST() \%>\% execute()
   cat(response$content)
}

}

