% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{FETCH}
\alias{FETCH}
\title{FETCH}
\usage{
FETCH(imap, msg_id, section = "HEADER")
}
\arguments{
\item{imap}{imap object}

\item{msg_id}{message id as returned
from \link{SEARCH} command}

\item{section}{section to fetch\cr
\code{""} - complete message\cr
\code{"HEADER"} - header\cr
\code{"TEXT"} - body\cr}
}
\value{
character
}
\description{
Fetch message (section) for given id.
}
\examples{
\dontrun{
   imap \%>\% SELECT("INBOX") \%>\%
      FETCH(1, "") \%>\% execute()
}


}

