% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{LIST}
\alias{LIST}
\title{LIST}
\usage{
LIST(imap, folder = "")
}
\arguments{
\item{imap}{imap object}

\item{folder}{mailbox name (empty name lists root folder)}
}
\value{
updated imap object
}
\description{
List folder.
}
\examples{
\dontrun{
   # with 'imap' as an imap object ...
   response <- imap \%>\% LIST() \%>\% execute()
}

}

