% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py-parse.R
\name{py_parse_header}
\alias{py_parse_header}
\title{Parse email header.}
\usage{
py_parse_header(filename)
}
\arguments{
\item{filename}{filename of raw message}
}
\value{
list with (to, from, date, subject)
}
\description{
Parses email header using python.
}
\examples{
\dontrun{
   imap <- IMAP("server") \%>\% user("uid", "pwd") \%>\%
      SELECT("INBOX")
   imap \%>\% FETCH(1) \%>\% execute(tempfile()) \%>\%
      py_parse_header()
}

}

