% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{SEARCH}
\alias{SEARCH}
\title{SEARCH}
\usage{
SEARCH(imap, criteria = "ALL")
}
\arguments{
\item{imap}{imap object}

\item{criteria}{search criteria like 'ALL', 'UNSEEN', 'FROM "stefan"'
et cetera}
}
\value{
updated imap object
}
\description{
Search mailbox for messages matching criteria.
}
\examples{
\dontrun{
   # with 'imap' as an imap object ...
   response <- imap \%>\%
      SEARCH("UNSEEN") \%>\% execute()
}
}

