% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{sections}
\alias{sections}
\title{List section options.}
\usage{
sections()
}
\value{
list of section options
}
\description{
List section options.
}
\examples{
sections()$header_min
}
\seealso{
\code{\link{FETCH}}
}

