% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imap.R
\name{user}
\alias{user}
\title{user}
\usage{
user(imap, username, password)
}
\arguments{
\item{imap}{imap object}

\item{username}{username}

\item{password}{password}
}
\value{
updated imap object
}
\description{
Set username and password needed to log in to imap server.
}
\examples{
imap <- IMAP("imaps://imap.gmail.com") \%>\%
   user("name", "pwd")

}

