% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute}
\alias{execute}
\title{Execute imap request.}
\usage{
execute(imap, filename = NULL, ...)
}
\arguments{
\item{imap}{imap object}

\item{filename}{if given, response will be written to disk}

\item{...}{optional parameters passed to curl,
see \code{\link[curl]{curl_options}}}
}
\value{
content OR filename
}
\description{
Execute imap request.
}
\examples{
\dontrun{
   imap <- IMAP("imaps://imap.gmail.com") \%>\%
      user("name", "pwd")

   imap \%>\% LIST() \%>\% execute() \%>\% cat()

   filename <- imap \%>\% SELECT("INBOX") \%>\%
      FETCH(1) \%>\% execute(tempfile())

   imap \%>\% LIST("INBOX") \%>\%
      execute(ssl_verifypeer = FALSE)
}

}

