% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{LIST}
\alias{LIST}
\title{LIST}
\usage{
LIST(imap)
}
\arguments{
\item{imap}{imap object}
}
\value{
updated imap object
}
\description{
list folders on server
}
\examples{
\dontrun{
   # when 'imap' is an imap object ...
   response <- imap \%>\% LIST() \%>\% execute()
}
}

