% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{SEARCH}
\alias{SEARCH}
\title{SEARCH}
\usage{
SEARCH(imap, request = "ALL")
}
\arguments{
\item{imap}{imap object}

\item{request}{search request like 'ALL', 'UNSEEN', ...}
}
\value{
updated imap object
}
\description{
Search mailbox for messages matching request.
}
\examples{
\dontrun{
   # with 'imap' as an imap object ...
   response <- imap \%>\%
      SEARCH("UNSEEN") \%>\% execute()
}
}

