% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{sections}
\alias{sections}
\title{List section options.}
\usage{
sections(select = NULL)
}
\arguments{
\item{select}{section to be returned}
}
\value{
list of section options
}
\description{
List section options.
}
\examples{
# show sections
sections()

# select section
sections("header_min")

}
\seealso{
\code{\link{FETCH}}
}

