% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{sections}
\alias{sections}
\title{List message sections.}
\usage{
sections(select = NULL)
}
\arguments{
\item{select}{only return selected section}
}
\value{
list of all sections OR selected section
}
\description{
List message sections.
}
\examples{
# show all (available) sections
sections()

# select section
sections("header_min")

}
\seealso{
\code{\link{FETCH}}
}

