% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{SELECT}
\alias{SELECT}
\title{SELECT}
\usage{
SELECT(imap, folder)
}
\arguments{
\item{imap}{imap object}

\item{folder}{mailbox/folder name}
}
\value{
updated imap object
}
\description{
Select mailbox/folder.
}
\examples{
imap <- IMAP("imaps://imap.gmail.com") \%>\% SELECT("INBOX")

}

